fixed bugs

+++ src/ArrayListMultiSet.java
@@ -24,7 +24,7 @@ public class ArrayListMultiSet extends MultiSet {
     @Override
     void remove(int item) {
         if (this.lst.contains(item)) {
-            this.lst.remove(item);
+            this.lst.remove((Integer) item);
         }
     }
